Unable to create swagger file due to memory explosion
For really large Yang models, Yanger went out of memory when
trying to generate the Swagger JSON output.

The solution consisted of two things:
1. Add the compressed option to one of the created ETS tables.
2. Write the generated output to the result file as soon as possible,
   i.e do not wait for the ets:foldl to gather all output, instead
   write output to file for each foldl iteration.

Also, changing the behaviour of the Swagger plugin to not allow errors,
but to allow warnings. The client can prohibit output to be generated
when there are warnings by using the -E switch.