Adjust augmented xpath-expressions.
For an augmenting model with an augment like:

  augment "/foo:routing-instances/foo:routing-instance" {
    when "/if:interfaces/if:interface[if:name=current()/foo:interfaces/foo:interface/foo:name]/if:type = 'access'";
    leaf default-access {
      type boolean;
    }
  }

Yanger compiled this and checks the xpath which is ok, then the
when-expression will be "moved into" the default-access leaf.  But
this means the XPATH-expression is no longer correct, because the
function "current()":s context-node has now moved one step.

To fix this we walk through the xpath-expression and correct *all*
"current()" by inserting apropriate number of parent-steps ("../").