JSTree plugin
Adds the jstree yanger plugin

Similar functionality to the corresponding plugin in pyang

There are a total of three options

The first option, jstree-help, will print the plugin's help function into the terminal.

The second option, jstree-server, is a mechanism for serving the generated HTML locally over HTTP.
This was a feature that was requested for in the ENG-26407 ticket. This option was not implemented in
the pyang version.

The third option, jstree-server-imp-mod, is also a mechanism for serving the generated HTML.
It  also generates html for a module's imported modules.
These imported modules can be accessed from a select on the generated webpage.