minor bugfix for prefix instead of modulename from imported yangs
Added yin-test for tailf-info meta-tags.

NOTE: Another issue arose: the regexp is being un-escaped
in the Erlang-version.

TODO: Find out if the original yang-file is wrong or if this
is expected behaviour. For now we are assuming that the yin-file
generated by pyang should be equal to the yanger yin-file.